data_generation/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ensure==1.0.4
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.7
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.1
